pyproject.toml
[tool.poetry]
name = "fast-tmp"
version = "0.1.24"
description = ""
authors = ["Chise1 <chise123@live.com>"]
include = ["LICENSE", "README.md"]
license = "Apache-2.0"
packages = [
    { include = "fastapi_cli" },
    { include = "fast_tmp" }
]
repository = "https://github.com/Chise1/fast-tmp"

[tool.poetry.dependencies]
python = "^3.7"
cookiecutter = "^1.7.2"
typer = "^0.3.2"
sqlalchemy = { extras = ["mypy"], version = "^1.4.31" }

[tool.poetry.dev-dependencies]
python-dotenv = "^0.15.0"
fastapi = "^0.73.0"
flake8 = "*"
isort = "*"
black = "^20.8b1"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-asyncio = "*"
asynctest = "*"
bandit = "*"
mypy = "*"
python-jose = "^3.2.0"
python-cas = "^1.5.0"
sentry-sdk = "^1.0.0"
aioredis = "^1.3.1"
aiofiles = "^0.8.0"
uvicorn = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
url = "https://mirrors.aliyun.com/pypi/simple/"
name = "aliyun"
default = true


[tool.poetry.scripts]
fast-tmp = "fastapi_cli:main"

[mypy]
plugins = "sqlalchemy.ext.mypy.plugin"
